Update config-options.md
I would add TRANSMISSION_WATCH_DIR on a new row, but given I don't use this option I'm not too familiar with what it does. Reason for adding this section is that this information is important when using other servarr docker containers. While bugfixing I read through some old issue pages that - TRANSMISSION_COMPLETED_DIRECTORY was a option whereas on current version and probably older the older versions as well was not a valid option. I would probably also add somewhere on the "Configuration options" page that these environment variables are fetched from https://github.com/haugene/docker-transmission-openvpn/blob/master/Dockerfile which are the most up to date variables. Thanks for your work on this project it is awesome!